# backend/swagger.yaml
openapi: 3.0.3
info:
  title: DOT Platform API
  description: |
    DOT 플랫폼 REST API 문서

    식음료 사업장을 위한 종합 운영 관리 시스템 API입니다.
    근태관리, 스케줄링, 급여관리 기능을 제공합니다.

    ## 인증
    모든 API는 JWT 토큰 기반 인증을 사용합니다.
    로그인 후 받은 토큰을 Authorization 헤더에 포함시켜 요청하세요.

    ## Rate Limiting
    - 인증된 사용자: 분당 100회
    - 비인증 사용자: 분당 20회

    ## 에러 코드
    - 400: 잘못된 요청
    - 401: 인증 실패
    - 403: 권한 없음
    - 404: 리소스 없음
    - 429: 요청 한도 초과
    - 500: 서버 오류

  version: 0.1.0
  contact:
    name: DOT Platform Support
    email: support@dotplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/v1
    description: 개발 서버
  - url: https://api.dotplatform.com/v1
    description: 프로덕션 서버
...
security:
  - bearerAuth: []
paths:
...
  /schedules:
    get:
      tags: [Schedules]
      summary: 스케줄 목록 조회
      description: 내 스케줄 또는 팀 스케줄 조회
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2024-01-15"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2024-01-21"
        - name: type
          in: query
          schema:
            type: string
            enum: [personal, team]
            default: personal
            example: "personal"
      responses:
        200:
          description: 스케줄 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'

    post:
      tags: [Schedules]
      summary: 스케줄 생성
      description: 새로운 스케줄 생성 (Manager 이상)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreate'
      responses:
        201:
          description: 스케줄 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "스케줄이 생성되었습니다"
                  data:
                    type: object
                    properties:
                      scheduleId:
                        type: integer
                        example: 789
  /schedules/{scheduleId}:
    parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          example: 789

    get:
      tags: [Schedules]
      summary: 스케줄 상세 조회
      description: 특정 스케줄의 상세 정보 조회
      responses:
        200:
          description: 스케줄 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Schedule'

    put:
      tags: [Schedules]
      summary: 스케줄 수정
      description: 스케줄 정보 수정 (Manager 이상)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpdate'
      responses:
        200:
          description: 스케줄 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "스케줄이 수정되었습니다"

    delete:
      tags: [Schedules]
      summary: 스케줄 삭제
      description: 스케줄 삭제 (Manager 이상)
      responses:
        200:
          description: 스케줄 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "스케줄이 삭제되었습니다"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer {token}'

  # ============================================
  # Schemas
  # ============================================
  schemas:
...
    Schedule:
      type: object
      properties:
        id:
          type: integer
          example: 789
        businessId:
          type: integer
          example: 456
        userId:
          type: integer
          example: 123
        date:
          type: string
          format: date
          example: "2024-01-15"
        startTime:
          type: string
          format: time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          example: "18:00:00"
        breakMinutes:
          type: integer
          example: 60
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
          example: "scheduled"
        notes:
          type: string
          nullable: true
          example: "오전 회의 참석 필요"
    ScheduleCreate:
      type: object
      required:
        - userId
        - date
        - startTime
        - endTime
      properties:
        userId:
          type: integer
          example: 123
        date:
          type: string
          format: date
          example: "2024-01-15"
        startTime:
          type: string
          format: time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          example: "18:00:00"
        breakMinutes:
          type: integer
          default: 60
          example: 60
        isRecurring:
          type: boolean
          default: false
          example: false
        recurrencePattern:
          type: string
          enum: [daily, weekly, monthly]
          example: "weekly"
        recurrenceEndDate:
          type: string
          format: date
          example: "2024-12-31"
    ScheduleUpdate:
      type: object
      properties:
        startTime:
          type: string
          format: time
          example: "10:00:00"
        endTime:
          type: string
          format: time
          example: "19:00:00"
        breakMinutes:
          type: integer
          example: 60
        notes:
          type: string
          example: "시간 변경됨"
...